requirements.txt
dj-database-url==0.5.0
Django==2.2.6
django-localflavor==2.2
django-model-utils==3.2.0
python-decouple==3.1
python-stdnum==1.12
pytz==2019.3
sqlparse==0.3.0
